refactor peer connections
