+++ minimint-api/src/lib.rs
@@ -17,6 +17,7 @@ pub mod config;
 pub mod db;
 pub mod encoding;
 pub mod module;
+pub mod net;
 pub mod rand;
 pub mod task;
