+++ minimint/src/lib.rs
@@ -1,6 +1,7 @@
 extern crate minimint_api;
 
 use std::collections::BTreeMap;
+use std::collections::BTreeSet;
 use std::future::Future;
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -16,18 +17,18 @@ use tracing::warn;
 
 use config::ServerConfig;
 use minimint_api::db::Database;
+use minimint_api::net::peers::{AnyPeerConnections, PeerConnections};
 use minimint_api::PeerId;
 use minimint_core::modules::ln::LightningModule;
 use minimint_core::modules::wallet::bitcoind::BitcoindRpc;
 use minimint_core::modules::wallet::{bitcoincore_rpc, Wallet};
 
+use crate::net::peers::PeerSlice;
 pub use minimint_core::*;
 
 use crate::consensus::{ConsensusItem, ConsensusOutcome, ConsensusProposal, MinimintConsensus};
 use crate::net::connect::{Connector, TlsTcpConnector};
-use crate::net::peers::{
-    AnyPeerConnections, PeerConnections, PeerConnector, ReconnectPeerConnections,
-};
+use crate::net::peers::{PeerConnector, ReconnectPeerConnections};
 use crate::rng::RngGenerator;
 
 /// The actual implementation of the federated mint
@@ -52,6 +53,7 @@ pub struct MinimintServer {
     pub connections: AnyPeerConnections<Message<PeerId>>,
     pub cfg: ServerConfig,
     pub hbbft: HoneyBadger<Vec<ConsensusItem>, PeerId>,
+    pub peers: BTreeSet<PeerId>,
 }
 
 /// Start all the components of the mint and plug them together
@@ -127,6 +129,7 @@ impl MinimintServer {
             hbbft,
             consensus,
             cfg: cfg.clone(),
+            peers: cfg.peers.keys().cloned().collect(),
         }
     }
 
@@ -193,7 +196,9 @@ impl MinimintServer {
             .expect("HBBFT propose failed");
 
         for msg in step.messages {
-            self.connections.send(msg.target, msg.message).await;
+            self.connections
+                .send(&msg.target.peers(&self.peers), msg.message)
+                .await;
         }
 
         while outcomes.is_empty() {
@@ -225,7 +230,9 @@ impl MinimintServer {
         }
 
         for msg in step.messages {
-            self.connections.send(msg.target, msg.message).await;
+            self.connections
+                .send(&msg.target.peers(&self.peers), msg.message)
+                .await;
         }
 
         step.output
